Accept smaller codepoints in each range
Eg. consider codepoints that are represented with exactly 2 octets
in UTF-8; those consume 11 bits from the address. These are the
codepoints between 2 ** 7 (inclusive) and 2 ** 11 (exclusive). The
current code rejects the codepoints between 2 ** 7 and 2 ** 8. In
addition, it would output null octets (since, interestingly, the
following does not fail: chr(0).encode('utf-8')).